More scaling stuff
Added buttons for scaling and limits to scaling with the buttons. Trying
to add limits (zoom in no further than scale(1,1) and out no further
than...scale(.125, .125)?)
to the mouse wheel scale feature doesn't work so well.
Unfortunately, the fact that there are no limits on the wheel scale
means it can mess up the counter on the button scale. Le sigh.